pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-vanguard"
version = "2.0.0-rc1"
description = "Security proxy and verification layer for the Model Context Protocol (MCP)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "provnai", email = "research@provnai.com" }]
keywords = ["mcp", "model-context-protocol", "ai-security", "railway", "firewall", "security-proxy", "vex-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.27",
    "google-re2>=1.1.20240701",
    "cryptography>=46.0.0",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "starlette>=0.30.0",
    "uvicorn>=0.23.0",
    "mcp>=1.0.0",
    "fastmcp>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
full = [
    "uvloop>=0.19; sys_platform != 'win32'",
]

[project.scripts]
vanguard = "core.cli:app"

[project.urls]
Homepage = "https://provnai.com"
Documentation = "https://provnai.dev"
Repository = "https://github.com/provnai/McpVanguard"
Issues = "https://github.com/provnai/McpVanguard/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:websockets.*:",
    "ignore::DeprecationWarning:uvicorn.*:",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "rules*", "assets*"]

[tool.setuptools.package-data]
rules = ["*.yaml", "*.json"]
assets = ["*.png"]
